requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
google-generativeai==0.3.2
aiohttp==3.9.1
replicate==0.22.0
openai>=1.0.0
SQLAlchemy==2.0.30
greenlet==3.2.2
aiosqlite==0.20.0 
python-telegram-bot[socks]==20.7
httpx[socks]